==> scripts/19_sem_results_further_analysis.R <==
# Get SEM path coefficient estimates table, etc

wd <- dirname(rstudioapi::getSourceEditorContext()$path)
setwd(wd)
library(tidyr)
library(ggplot2)
theme_set(theme_bw())
library(cowplot)
library(sf)
source("0_functions.R")
library(spdep)
library(lavaan)
library(png)
library(Simpsons)
library(DiagrammeR)
library(ggsci)
library(plyr)
rm(list = setdiff(ls(), lsf.str()))  


# Conceptual model figure -------------------------------------------------

tmp = 
grViz("
digraph concept {
  # Node statements
      node [shape = Mrecord, fontname = Helvetica, fontsize=10, style=filled, 
          color=white, peripheries=2, fillcolor=grey90]
      SR; MRD   
    
      node [shape = ellipse, style=filled, fillcolor = '#71A8F4', 
            color=white, peripheries=1, margin=0.05]
      mat; pre

      node [shape = oval, style=filled, fillcolor = '#F4F471', 
            color=white, peripheries=1]
      prs; tra

      node [shape = oval, style=filled, fillcolor = '#F48771', 
            color=white, peripheries=1]
      mat_ano_mio[label='Miocene \n anomaly']
      mat_ano_lgm[label='LGM mat \n anomaly']

    node [shape = oval, style=filled, fillcolor = '#9DF471', 
          color=white, peripheries=1, rank=top]
      sub_trop_mbf[label='(sub)trop \n mbf']
      temp_mbf[label='temp. \n bmf']
      sub_trop_dbf[label='(sub)trop \n dbf']
      mont_gs[label='mont. gs']
    
    node [shape = oval, style=filled, fillcolor = '#F471B2', 
          color=white, peripheries=1, rank=a]
    soil;    area;    tri;    elev_range[label='elev.\nrange']
    sd_tra[label='sd(tra)']
    sd_prs[label='sd(prs)']
    sd_mat[label='sd(mat)']
    sd_pre[label='sd(pre)']

  # Edge statements
{sd_pre; elev_range; mat_ano_lgm; mont_gs; mat; sd_mat; sd_prs; tra; soil; sub_trop_mbf ; area; MRD} -> SR[style=fill]
sub_trop_dbf -> SR[len=0.7, col=black]
{mat_ano_mio; elev_range; sd_prs; soil; tra; sd_tra; tri; sd_pre; temp_mbf; pre; sub_trop_mbf; prs} -> MRD[len='', style=fill, color='grey50']
area -> {sd_prs; sd_tra; sd_mat; sd_pre; soil}[len='', color='#F471B2']
{pre; tra; mat} -> sub_trop_mbf[len='', color = '#9DF471']

 # Arranging edge statements
  edge[style=invis]
  sub_trop_dbf -> sub_trop_mbf
  temp_mbf -> sub_trop_mbf
  mont_gs -> sub_trop_mbf[weight=2, len=0.2]
  pre -> mat
  soil -> sd_pre
  tra -> prs[weight=2, len=0.2]
  mat_ano_lgm -> mat_ano_mio
  

//    graph[nodesep=0.1, layout=dot, ranksep=0.1, splines=true, compound=true]
//    graph[nodesep=0.01, layout=circo, mindist=.2]  
  graph[layout=neato, mindist=0.5, root=SR, splines=true, overlap=false]
}
")
tmp = DiagrammeRsvg::export_svg(tmp)
tmp = charToRaw(tmp) # flatten
#rsvg::rsvg_svg(tmp, "../figures/concept_figure.svg")




# Table of estimates ###########################################################################
rm(list = setdiff(ls(), lsf.str()))  
load("../processed_data/best_model.RDATA")
shp <- readRDS("../processed_data/shp_object_fin_analysis.RDS")
dat_no.na <- readRDS("../processed_data/sem_input_data.RDS")
dat_no.na$level3 <- row.names(dat_no.na)

# area included in analyis (load the prelim shp file including all areas)
#sum(shp$area[shp$LEVEL_3_CO %in% dat_no.na$level3]) / sum(shp$area[shp$LEVEL_3_CO!="ANT"])


## add indirect effect names to variables
cat(max.var.mod_mod7)
summary(max.var.mod.fit.mod7, standardized = TRUE, fit.measures=TRUE, rsq=TRUE)

fm <- "
sr_trans ~ s*soil + trf*sub_trop_mbf + a*area + mrd + ms*mont_gs + t*tra_m + m*mat_m + prs_m + tr0*tri
mrd ~ p2*pre_m + trf2*sub_trop_mbf + prs_m + mio_mat_ano_m + s1*soil + tri + a3*area 
soil ~ a1*area + ms1*mont_gs
sub_trop_mbf ~ p1*pre_m + t1*tra_m + m1*mat_m + a2*area + tr1*tri

# indirect paths on SR
sr_area_via_trf := a2*trf
sr_area_via_soil := a1*s
sr_area_total := a + a1*s + a2*trf

sr_pre_m_via_trf := p1*trf
sr_tri_via_trf := tr1*trf
sr_mont_gs_via_soil := ms1*s

sr_tra_m_via_trf := t1*trf
sr_tra_m_total := t1*trf + t

sr_mat_m_via_trf := m1*trf
sr_mat_m_total :=  m1*trf + m


# indirect paths on MRD
mrd_area_via_soil := a1*s1
mrd_area_total := a1*s1 + a3
mrd_pre_via_trf := p1*trf2
mrd_pre_total := p1*trf2 + p2
"


fm.fit <- sem(fm, data = dat_no.na, estimator="MLM")
summary(fm.fit, standardized = TRUE, rsq=T, fit.measures=TRUE)
fitmeasures(fm.fit, c("cfi.robust", "rmsea.robust", "aic", "pvalue.scaled", "chisq.scaled", "df", "chisq.scaling.factor"))

# documentation notes that standardized = TRUE only displays standardized estimates. "Note that SEs and tests are still based on unstandardized estimates. Use standardizedSolution() to obtain SEs and test statistics for standardized estimates"

pres2 <- standardizedSolution(fm.fit)
# filter for direct and indirect paths
pres2 <- pres2[pres2$op=="~" |pres2$op==":=" ,-grep("z|ci", names(pres2))]
# round estimates and p-values separately
pres2[,grep("est|se", names(pres2))] <- round(pres2[,grep("est|se", names(pres2))],2)
pres2[,grep("pvalue", names(pres2))] <- round(pres2[,grep("pvalue", names(pres2))],3)
temp <- pres2[order(pres2$op, pres2$lhs, abs(pres2$est.std), decreasing = T),]

#write.csv(temp,"../processed_data/sem/sem_model_output.csv") 



# Model path coefficients and correlations ############################################################
lavCor(fm.fit) # observed correlations
inspect(fm.fit, what="cor.all") # model-implied correlations among variables
resid(fm.fit, "cor")
# Large positive values indicate the model under-predicts the correlation; large negative values suggest over-prediction of correlation. Values around |r>.1| are worth closer consideration.


# Univariate VS model regressions ##############################################################
## Scatterplot showing all significant effects for SR and MRD
pres2 <- standardizedSolution(fm.fit)
pres3 <- pres2[pres2$pvalue<0.05 & pres2$op=="~",]
sig.sr <- pres3[pres3$lhs %in% c("sr_trans"),]
plot_sr <- dat_no.na[,c("sr_trans", sig.sr$rhs)]


plot_sr <- pivot_longer(plot_sr, col=-sr_trans)
plot_sr$std.all <- rep(sig.sr$est.std, nrow(dat_no.na))


scatterplot_sr <- ggplot(plot_sr, aes(x=value, y=sr_trans))+
  geom_point(alpha=0.3)+
  ylab("species richness")+
  facet_wrap(~name, scales = "free", ncol = 4)+# labeller = as_labeller(var_names_sr))+
  geom_smooth(method="lm", se=F)+
  geom_abline(aes(slope=std.all, intercept=0), col="grey20")+
  theme(strip.background = element_blank())
#ggsave("../figures/scatterplot_sig_effects_SR.png",  width=7, height=7, units = "in", dpi = 600)

sig.mrd <- pres3[pres3$lhs %in% c("mrd"),]
plot_mrd <- dat_no.na[,c("mrd", sig.mrd$rhs)]
plot_mrd <- pivot_longer(plot_mrd, col=-mrd)
plot_mrd$std.all <- rep(sig.mrd$est.std, nrow(dat_no.na))
scatterplot_mrd <- ggplot(plot_mrd, aes(x=value, y=mrd))+
  geom_point(alpha=0.3)+
  ylab("mean root distance")+
  facet_wrap(~name, scales = "free", ncol=4)+
  geom_smooth(method="lm", se=F)+
  geom_abline(aes(slope=std.all, intercept=0), col="grey20")+
  theme(strip.background = element_blank())
#ggsave("figures/scatterplot_sig_effects_MRD.png",  width=5, height=5, units = "in", dpi = 600)

plot_grid(ncol = 1, rel_heights = c(1, 1),
          labels = c("A", "B"),  label_fontface = "plain",
          scatterplot_sr, scatterplot_mrd)
#ggsave("../figures/scatterplot_sig_effects_all.png",  width=8, height=8, units = "in", dpi = 600)



# Interaction effects / Simpsons paradox ###########

## SR tra_m and tra_m 
cor(dat_no.na[,c("sr_trans", "mat_m", "prs_m", "tra_m")])
n=10
dat_no.na$mat_breaks <- c(1:n)[cut(dat_no.na$mat_m, breaks=n)]
# dat_no.na$tra_breaks <- c(1:n)[cut(dat_no.na$tra_m, breaks=n)]
# dat_no.na$prs_breaks <- c(1:n)[cut(dat_no.na$prs_m, breaks=n)]

simp_tra1<- Simpsons(tra_m, sr_trans, data=dat_no.na, nreps=500) 
simp_tra1<- Simpsons(tra_m, sr_trans, clusterid = mat_breaks, data=dat_no.na, nreps=500)
simp_sea1<-Simpsons(prs_m, sr_trans, data=dat_no.na, nreps=500) 
simp_sea1<-Simpsons(prs_m, sr_trans, clusterid = mat_breaks, data=dat_no.na, nreps=500) 
# signs for simpsons paradox in both when clustering for mat_m



tra_sr <- ggplot(dat_no.na, aes(x=tra_m, y=sr_trans, col=factor(mat_breaks)))+
    geom_point()+
    scale_color_viridis_d(option = "plasma", 'MAT bins')+
    #scale_color_gradient2('MAT intervals', low = "blue", mid = "yellow", high = "red", midpoint = 5)
    #scale_color_continuous("MAT intervals")+
    guides(colour = guide_legend(reverse=T))+
    geom_smooth(method = "lm", se=F, col="grey")+
    geom_smooth(method = "lm", aes(group=mat_breaks), se=F)+
    theme(legend.position = "none")+
    ylab("species richness")
prs_sr <- ggplot(dat_no.na, aes(x=prs_m, y=sr_trans, col=factor(mat_breaks)))+
    geom_point()+
    scale_color_viridis_d(option = "plasma", 'MAT bins')+
    guides(colour = guide_legend(reverse=T))+
    geom_smooth(method = "lm", se=F, col="grey")+
    geom_smooth(method = "lm", aes(group=mat_breaks), se=F)+
    ylab("species richness")


# ## MRD area
# n=10
# dat_no.na$soil_breaks <- c(1:n)[cut(dat_no.na$soil, breaks=n)]
# table(dat_no.na$soil_breaks)
# 
# simp_area1 <- Simpsons(area, mrd, data=dat_no.na, nreps=500) 
# simp_area1 <- Simpsons(area, mrd, clusterid = soil_breaks, data=dat_no.na, nreps=500) 
# 
# area_mrd <- ggplot(dat_no.na, aes(x=area+abs(min(area)), y=mrd, col=factor(soil_breaks)))+ 
#   geom_point()+
#   scale_x_continuous("area", trans="sqrt")+
#   scale_color_viridis_d(option = "plasma", 'soil bins')+
#   geom_smooth(method = "lm", se=F, col="grey")+
#   geom_smooth(method = "lm", aes(group=soil_breaks), se=F)+
#   guides(colour = guide_legend(reverse=T))#+
# empty <- ggplot(dat_no.na, aes(x=area+abs(min(area)), y=mrd, col=factor(soil_breaks)))+ 
#   geom_blank()+
#   theme_void()

top_row <- plot_grid(ncol = 2, rel_widths = c(0.8,1),
          labels = c("A", "B"), label_fontface = "plain",
          tra_sr, prs_sr)
# bottom_row <- plot_grid(ncol = 2, rel_widths = c(1, 0.77),
#                         labels = c("", ""),
#                         area_mrd, empty)
plot_grid(top_row, nrow=1)
#ggsave("../figures/seasonality_paradox.png", width=8, height=4, units = "in", dpi = 600)


## Global patterns for temperature range influence on SR
### get slopes and p-values for each mat bin
(res <- t(sapply(split(dat_no.na, list(dat_no.na$mat_breaks)),
         function(subd){summary(lm(sr_trans ~ tra_m, data = subd))$coefficients[2,c(1,4)]}))) 
res <- as.data.frame(res)
names(res) <- c("tra_coef", "tra_pvalue")

# get mean mat_m per group
mat_mean <- tapply(dat_no.na$mat_m, dat_no.na$mat_breaks, mean)
res$mat_bin_means <- as.numeric(mat_mean)

ggplot(res, aes(x=mat_bin_means, y=tra_coef))+
  geom_point(aes(shape=tra_pvalue<0.05, col=tra_coef), size=1.7)+
  #scale_color())+
  scale_y_continuous("SR~tra coefficient")+
  theme(legend.position = "null", 
        plot.background = element_blank(),
    panel.background = element_blank())+
  geom_hline(yintercept = 0, lty=2)
ggsave("../figures/slope_mat.png",width=2.5, height=2,units="in")

# visualize as map
simpsons_map <- shp[,c("LEVEL_3_CO", "sr", "mrd", "geometry")]
simpsons_map <- st_transform(simpsons_map, "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs")
simpsons_map$tra_dynamics <- "positive"
simpsons_map$tra_dynamics[simpsons_map$LEVEL_3_CO %in% 
                            dat_no.na$level3[dat_no.na$mat_breaks %in% c(9,10)]] <- "negative"
# add slopes for all bins
res$mat_breaks <- c(1:10)
# merge level 3 in via mat_bin_means
temp <- merge(dat_no.na[, c("mat_breaks", "level3")], res, all.x=TRUE)
simpsons_map <- merge(simpsons_map, temp,
                      by.x="LEVEL_3_CO", by.y="level3", all.x=TRUE)

simpsons_map$tra_coef2 <- simpsons_map$tra_coef
simpsons_map$tra_coef2[simpsons_map$tra_pvalue>0.05] <- NA
simpsons_map <- simpsons_map[simpsons_map$LEVEL_3_CO!="ANT",]

(my_plot3 <- ggplot(simpsons_map) + 
  geom_sf(aes(fill = tra_coef2), lwd=0.1) + 
 scale_fill_continuous("", na.value="grey70")+
  theme_void()+
theme(legend.position = c(0.265, 0.322), # c(0.265, 0.322) excluding ANT
      legend.background = element_blank(),
      legend.key = element_blank(),
      legend.key.width = unit(3, "mm"),
      legend.key.height = unit(6.5, "mm"),
      legend.text = element_blank())
)

logo_file <- readPNG("../figures/slope_mat.png")
my_plot_4 <- ggdraw() +
  draw_image(logo_file,  x = -0.35, y = -0.14, scale = .25) + #x = -0.35, y = -0.14, scale = .25
  draw_plot(my_plot3)
my_plot_4
ggsave("../figures/tra_scale_map.png", dpi=600, width=10, height=7, bg="white")


## Global patterns for precipitation seasonality influence on SR
### get slopes and p-values for each mat bin
(res <- t(sapply(split(dat_no.na, list(dat_no.na$mat_breaks)),
                 function(subd){summary(lm(sr_trans ~ prs_m, data = subd))$coefficients[2,c(1,4)]}))) 
res <- as.data.frame(res)
names(res) <- c("prs_coef", "prs_pvalue")

# get mean mat_m per group
#mat_mean <- tapply(dat_no.na$mat_m, dat_no.na$mat_breaks, mean)
res$mat_bin_means <- as.numeric(mat_mean)

ggplot(res, aes(x=mat_bin_means, y=prs_coef))+
  geom_point(aes(shape=prs_pvalue<0.05, col=prs_coef), size=1.7)+
  #scale_color())+
  scale_y_continuous("SR~prs coefficient")+
  theme(legend.position = "null", 
        plot.background = element_blank(),
        panel.background = element_blank())+
  geom_hline(yintercept = 0, lty=2)
ggsave("../figures/slope_mat2.png",width=2.5, height=2,units="in")

# visualize as map
res$mat_breaks <- c(1:10)
# merge level 3 in via mat_bin_means
temp <- merge(dat_no.na[, c("mat_breaks", "level3")], res, all.x=TRUE)
simpsons_map <- merge(simpsons_map, temp[,c("level3","prs_coef","prs_pvalue")],
                      by.x="LEVEL_3_CO", by.y="level3", all.x=TRUE)

simpsons_map$prs_coef2 <- simpsons_map$prs_coef
simpsons_map$prs_coef2[simpsons_map$prs_pvalue>0.05] <- NA
(my_plot3 <- ggplot(simpsons_map) + 
    geom_sf(aes(fill = prs_coef2), lwd=0.1) + 
    scale_fill_continuous("", na.value="grey70", 
                          limits=c(min(simpsons_map$prs_coef, na.rm=T),
                                  max(simpsons_map$prs_coef, na.rm=T)))+
    theme_void()+
    theme(legend.position = c(0.265, 0.383), # c(0.265, 0.322) excluding ANT
          legend.background = element_blank(),
          legend.key = element_blank(),
          legend.key.width = unit(3, "mm"),
          legend.key.height = unit(6, "mm"),
          legend.text = element_blank())
)

logo_file <- readPNG("figures/slope_mat2.png")
my_plot_4 <- ggdraw() +
  draw_image(logo_file,  x = -0.35, y = -0.12, scale = .25) + #x = -0.35, y = -0.14, scale = .25
  draw_plot(my_plot3)
my_plot_4
ggsave("../figures/tra_scale_map.png", dpi=600, width=10, height=7)






# Spatial autocorrelation ########################################################################
# Uses code from https://github.com/jebyrnes/spatial_correction_lavaan to get
# residuals

dat_no.na <- dat_no.na[order(dat_no.na$level3),]
load("../processed_data/best_model.RDATA")
# **** Table of estimates ----------------------------------------------
## add indirect effect names to variables
cat(max.var.mod_mod7)

# get coordinates 
# fix sf trouble...
sf::sf_use_s2(FALSE) 
shp$centroids <- st_centroid(shp) %>% 
  st_coordinates()
shp$y <- shp$centroids[,1] # x=lng
shp$x <- shp$centroids[,2]

# subset sf object to regions included in SEM
shp <- shp[shp$LEVEL_3_CO %in% dat_no.na$level3,]
shp <- merge(shp, dat_no.na[,c("sr_trans", "level3")], by.x="LEVEL_3_CO", by.y="level3")

# Get model residuals
## Fit the model without indirect effects and meanstructure=TRUE as required for 
## residual function and without path names to for later correction
fm.npn <- max.var.mod_mod7

fm.p <- sem(fm.npn, data = dat_no.na, estimator="MLM", meanstructure=TRUE)
resids <- as.data.frame(my.residuals_lavaan(fm.p)) 
rawdata <- data.frame(inspect(fm.p, "data")) # get raw data from the model
names(resids) <- paste0(names(resids),"_residuals")
plot(rawdata$sr_trans, dat_no.na$sr_trans) # check order for attaching level3 ID
resids$level3 <- dat_no.na$level3

# add residuals to spatial object
shp <- merge(shp, resids, by.x="LEVEL_3_CO", by.y="level3", all.x=TRUE) 

# add fitted SR values to spatial object
fitted_vals <- my.fitted_lavaan(fm.p)
names(fitted_vals) <- paste0(names(fitted_vals),"_fitted")
fitted_vals$level3 <- dat_no.na$level3 
shp <- merge(shp, fitted_vals[,c("sr_trans_fitted", "level3")], by.x="LEVEL_3_CO", by.y="level3", all.x=TRUE)

# plot original vs fitted and residuals
plot_grid(labels = c("A", "B", "C"), label_fontface = "plain", ncol = 2,
          ggplot(shp, aes(x=sr_trans, y=sr_trans_fitted))+
            geom_point()+
            xlab("species richness")+
            ylab("species richness fitted")+
            geom_abline(slope=1, intercept=0)
          ,
          ggplot(shp, aes(x=sr_trans, y=sr_trans_residuals))+
            geom_point()+
            xlab("species richness")+
            ylab("species richness SEM residuals")+
            geom_abline(slope=0, intercept=0)
          ,
          ggplot(shp, aes(x=abs(y), y=sr_trans_residuals, col=sr_trans))+
            geom_point()+
            ylab("species richness SEM residuals")+
            xlab("absolute latitude")+
            scale_color_continuous("SR")+
            geom_smooth(method="lm")
          )
#ggsave(file="../figures/SEM_residuals_SR.png",
#       width=7, height=7, units = "in", dpi = 600)



## Residuals spatial autocorrelation ###################################################

# Weighted distance matrix
distMat <- as.matrix(dist(cbind(shp$x, shp$y)))
distsInv <- 1/distMat # invert matrix for weights
diag(distsInv) <- 0

## GLOBAL autocorrelation
MI.sr <- moran.mc(shp$sr_trans_residuals[,1], mat2listw(distsInv, style = "B"), nsim=599, zero.policy=T)
MI.mrd  <-  moran.mc(shp$mrd_residuals[,1], mat2listw(distsInv, style = "B"), nsim=599, zero.policy=T)
MI.sr$statistic
MI.mrd$statistic

## DISTANCE bands
# distances are in kilometers
moran <- data.frame(dist.class = seq(100, 10000, 100),
                    sr.moransI = NA,
                    sr.moransp = NA,
                    mrd.moransI = NA,
                    mrd.moransp = NA)
coo <- cbind(shp$y, shp$x)
for(i in 1:length(moran$dist.class)){
  S.dist  <-  dnearneigh(coo, 0, moran$dist.class[i], longlat = TRUE)
  lw <- nb2listw(S.dist, style="W",zero.policy=T) 
  
  MI <- moran.mc(shp$sr_trans_residuals[,1], lw, nsim=599,zero.policy=T) 
  moran$sr.moransI[i] <- MI$statistic
  moran$sr.moransp[i] <- MI$p.value
  
  MI.mrd <- moran.mc(shp$mrd_residuals[,1], lw, nsim=599,zero.policy=T) 
  moran$mrd.moransI[i] <- MI.mrd$statistic
  moran$mrd.moransp[i] <- MI.mrd$p.value
  #moran$avg.n[i] <- median(card(lw$neighbours))
  if(!i%%1)cat(i,"\r")
}

temp <- pivot_longer(moran, cols = c("sr.moransI", "mrd.moransI"))
ggplot(temp, aes(x=dist.class, y=value, col=name)) +
  geom_line()+
  scale_x_continuous("Distance class (km)")+
  scale_color_discrete("SEM residuals", labels=c("MRD", "SR"))+
  ylab("Moran's I")
#ggsave("../figures/SAC_sem_residuals.png", width=5, height=4, units = "in", dpi = 600)


## Correct estimates for spatial autocorrelation

# merge coordinates into original model data
coo.df <- data.frame(level3=shp$LEVEL_3_CO, x=shp$x, y=shp$y)
dat_no.na <- merge(dat_no.na, coo.df, all.x=TRUE)

fm.fit.SA <- lavSpatialCorrect_correct(fm.p, xvar=dat_no.na$x, yvar=dat_no.na$y)
# observed autocorrelation matches global calculation

# compare to original
res <- rbind(
  round(fm.fit.SA$parameters$sr_trans[,c(2:6)],3),
  round(fm.fit.SA$parameters$mrd[,c(2:6)],3),
  round(fm.fit.SA$parameters$sub_trop_mbf[,c(2:6)],3),
  round(fm.fit.SA$parameters$soil[,c(2:6)],3)
)
# remove intercepts etc
res <- res[-which(grepl("~~|~1",row.names(res))),]
res$lhs <- gsub("~.*", "", row.names(res))
res$rhs <- gsub(".*~", "", row.names(res))
names(res)[1:5] <- paste0(names(res)[1:5],"_sac")

# merge into original estimate table
all <- merge(pres2, res)
all <- all[,c("lhs", "rhs", "est.std","se", "pvalue",
              "Estimate_sac", "Std.err_sac", "P(>|z|)_sac")]
temp <- all[order(all$lhs, abs(all$est.std), decreasing = T),]
temp$est.std <- round(temp$est.std,2)
temp$se <- round(temp$se,2)
temp$pvalue <- round(temp$pvalue,2)
temp$Estimate_sac <- round(temp$Estimate_sac,2)
temp$Std.err_sac <- round(temp$Std.err_sac,2)
#write.csv(temp, "../processed_data/sem_model_coefficients_SAC_correction.csv")







# Maps and  figures ########################################################


# **** SEM figure ----------------------------------------------------------

## ***** variable names version ####
#tmp = 
grViz("
digraph SEM {

  # Node statements
//  subgraph clustercenter{
    node [shape = box, fontname = Helvetica, fontsize=10, style=filled, 
    color=white, peripheries=2, fillcolor=grey90]
      SR[label='SR \n R&#xb2;=0.65']
      MRD [label='MRD \n R&#xb2;=0.56']
//  }
//  subgraph clusterclimate {
    node [shape = ellipse, style=filled, fillcolor = '#71A8F4', 
    color=white, peripheries=2, margin=0.05]
    mat
    pre
//  }
//  subgraph clusterbio {
    node [style=filled, fillcolor = '#9DF471', 
    color=white, peripheries=2, margin=0.01]
    sub_trop_mbf[label='(sub)trop bmf']
    mont_gs[label='mont g&s']
//  }
//  subgraph clusterseason {
    node [style=filled, fillcolor = '#F4F471', 
    color=white, peripheries=2, margin=0.05]
      prs; tra
//  }
//  subgraph clusterEH {
    node [style=filled, fillcolor = '#F471B2', 
    color=white, peripheries=2, margin=0.05]
    soil; area; tri
//  }
    node [style=filled, fillcolor = '#F48771', 
    color=white, peripheries=2]
    mat_ano[label='mat \n anomaly', margin=0.01]

  # Edge statements
  
  edge [color = red, penwidth=1, len=1]
    prs -> SR
    area -> MRD
  
  edge [color = red, penwidth=1, style=dashed, len=1]
    prs -> MRD
    tra -> sub_trop_mbf
  
  edge [color = red, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {pre sub_trop_mbf} -> MRD
    
  edge [color = black, penwidth=1, style=normal, len=1, arrowsize=1]
    {tra tri area} -> SR
    mat_ano -> MRD
    mont_gs -> soil
    {area mat tri}  -> sub_trop_mbf
  
  edge [color = black, penwidth=1, style=dashed, minlen=1]
    {mont_gs MRD} -> SR
    tri -> MRD
  
  edge [color = black, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {mat sub_trop_mbf} -> SR
    soil -> MRD
    area -> soil
  
  edge [color = black, penwidth=7, style=normal, arrowsize=0.4]
    soil -> SR
    pre -> sub_trop_mbf
    
  edge [style=invis]
    pre -> SR
    
  graph[layout=twopi, ranksep=1.7, root=SR, splines=true]
    
}
")
## ***** Full names version ####
#tmp = 
grViz("
digraph SEM {

  # Node statements
//  subgraph clustercenter{
    node [shape = box, fontname = Helvetica, fontsize=10, style=filled, 
    color=white, peripheries=2, fillcolor=grey90]
      SR[label='SR \n R&#xb2;=0.65']
      MRD [label='MRD \n R&#xb2;=0.56']
//  }
//  subgraph clusterclimate {
    node [shape = ellipse, style=filled, fillcolor = '#71A8F4', 
    color=white, peripheries=2, margin=0.05]
    mat[label='mean annual\n temperature']
    pre[label='total annual\n precipitation']
//  }
//  subgraph clusterbio {
    node [style=filled, fillcolor = '#9DF471', 
    color=white, peripheries=2]
    sub_trop_mbf[label='(sub)tropical moist\nbroadleaf forest']
    mont_gs[label='montane grass-\n & shrublands']
//  }
//  subgraph clusterseason {
    node [style=filled, fillcolor = '#F4F471', 
    color=white, peripheries=2]
      prs[label='precipitation \n seasonality']
      tra[label='annual \n temperature range']
//  }
//  subgraph clusterEH {
    node [style=filled, fillcolor = '#F471B2', 
    color=white, peripheries=2]
    soil[label='n soil types']
    area
    tri[label='terrain \n ruggedness']
//  }
    node [style=filled, fillcolor = '#F48771', 
    color=white, peripheries=2]
    mat_ano[label='Miocene \n temperature anomaly']

  # Edge statements
  
  edge [color = red, penwidth=1, len=1]
    prs -> SR
    area -> MRD
  
  edge [color = red, penwidth=1, style=dashed, len=1]
    prs -> MRD
    tra -> sub_trop_mbf
  
  edge [color = red, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {pre sub_trop_mbf} -> MRD
    
  edge [color = black, penwidth=1, style=normal, len=1, arrowsize=1]
    {tra tri area} -> SR
    mat_ano -> MRD
    mont_gs -> soil
    {area mat tri}  -> sub_trop_mbf
  
  edge [color = black, penwidth=1, style=dashed, minlen=1]
    {mont_gs MRD} -> SR
    tri -> MRD
  
  edge [color = black, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {mat sub_trop_mbf} -> SR
    soil -> MRD
    area -> soil
  
  edge [color = black, penwidth=5, style=normal, arrowsize=0.4]
    soil -> SR
    pre -> sub_trop_mbf
    
  edge [style=invis]
    pre -> SR
//    mat_ano -> SR

    
//  graph[nodesep=0.1, layout=dot, ranksep=.3, rankdir=BT, splines=true]
//  graph[nodesep=0.01, layout=circo, mindist=.7]  
  graph[layout=twopi, ranksep=1.5, root=SR, splines=true, overlap=false]
    
}
")
# 2. Convert to SVG, then save as png
tmp = DiagrammeRsvg::export_svg(tmp)
tmp = charToRaw(tmp) # flatten
#rsvg::rsvg_svg(tmp, "../figures/sem.svg")

## ***** separated bubbles version #####
# SR bubble ---------------------------------------------------------------

sr <- grViz("
digraph SEM {

  # Node statements
//  subgraph clustercenter{
    node [shape = circle, fontname = 'Helvetica', fontsize=10, style=filled, 
    color=white, peripheries=2, fillcolor=grey90, margin=0]
      SR[label='Species richness \n R&#xb2;=0.65']
//  }
//  subgraph clusterclimate {
    node [shape = rectangle, style=filled, fillcolor = '#71A8F4', 
    color=white, peripheries=2, margin=0.05]
    mat
    pre
//  }
//  subgraph clusterbio {
    node [style=filled, fillcolor = '#9DF471', 
    color=white, peripheries=2, margin=0.01]
    sub_trop_mbf[label='(sub)trop bmf']
    mont_gs[label='mont g&s']
//  }
//  subgraph clusterseason {
    node [style=filled, fillcolor = '#F4F471', 
    color=white, peripheries=2, margin=0.05]
      prs; tra
//  }
//  subgraph clusterEH {
    node [style=filled, fillcolor = '#F471B2', 
    color=white, peripheries=2, margin=0.05]
    soil; area; tri
//  }
    node [style=filled, fillcolor = '#F48771', 
    color=white, peripheries=2]
    mat_ano[label='mat \n anomaly', margin=0.01]

  # Edge statements
  
  edge [color = red, penwidth=1, len=1]
    prs -> SR

  edge [color = red, penwidth=1, style=dashed, len=1]
    tra -> sub_trop_mbf
  
  edge [color = red, penwidth=3, style=normal, len=1, arrowsize=0.7]
    
  edge [color = black, penwidth=1, style=normal, len=1, arrowsize=1]
    {tra tri area} -> SR
    mont_gs -> soil
    {area mat tri}  -> sub_trop_mbf
  
  edge [color = black, penwidth=1, style=dashed, minlen=1]
    {mont_gs} -> SR

  
  edge [color = black, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {mat sub_trop_mbf} -> SR

    area -> soil
  
  edge [color = black, penwidth=7, style=normal, arrowsize=0.4]
    soil -> SR
    pre -> sub_trop_mbf
    
  edge [style=invis]
    pre -> SR
    mat_ano -> SR
    
  graph[layout=twopi, ranksep=2, root=SR, splines=true]
    
}
")


# MRD bubble --------------------------------------------------------------

mrd <- grViz("
digraph SEM2 {

  # Node statements
//  subgraph clustercenter{
    node [shape = circle, fontname = Helvetica, fontsize=10, style=filled, 
    color=white, peripheries=3, fillcolor=grey90, margin=0]
      MRD [label='Diversification \n R&#xb2;=0.56', margin=0]
//  }
//  subgraph clusterclimate {
    node [shape = rectangle, style=filled, fillcolor = '#71A8F4', 
    color=white, peripheries=2, margin=0.05]
    mat
    pre
//  }
//  subgraph clusterbio {
    node [style=filled, fillcolor = '#9DF471', 
    color=white, peripheries=2, margin=0.01]
    sub_trop_mbf[label='(sub)trop bmf']
    mont_gs[label='mont g&s']
//  }
//  subgraph clusterseason {
    node [style=filled, fillcolor = '#F4F471', 
    color=white, peripheries=2, margin=0.05]
      prs; tra
//  }
//  subgraph clusterEH {
    node [style=filled, fillcolor = '#F471B2', 
    color=white, peripheries=2, margin=0.05]
    soil; area; tri
//  }
    node [style=filled, fillcolor = '#F48771', 
    color=white, peripheries=2]
    mat_ano[label='mat \n anomaly', margin=0.01]

  # Edge statements
  
  edge [color = red, penwidth=1, len=1]
    area -> MRD
  
  edge [color = red, penwidth=1, style=dashed, len=1]
    prs -> MRD
    tra -> sub_trop_mbf
  
  edge [color = red, penwidth=3, style=normal, len=1, arrowsize=0.7]
    {pre sub_trop_mbf} -> MRD
    
  edge [color = black, penwidth=1, style=normal, len=1, arrowsize=1]
    mat_ano -> MRD
    mont_gs -> soil
    {area mat tri}  -> sub_trop_mbf
  
  edge [color = black, penwidth=1, style=dashed, minlen=1]
    tri -> MRD
  
  edge [color = black, penwidth=3, style=normal, len=1, arrowsize=0.7]
    soil -> MRD
    area -> soil
  
  edge [color = black, penwidth=7, style=normal, arrowsize=0.4]

    pre -> sub_trop_mbf
    
  edge [style=invis]
  {tra mat mont_gs} -> MRD
    
  graph[layout=twopi, ranksep=2, root=MRD, splines=true]
    
}
")

# 2. Convert to SVG, then save as png
sr = DiagrammeRsvg::export_svg(sr)
sr = charToRaw(sr) # flatten
rsvg::rsvg_svg(sr, "../figures+tables/sr_bubble.svg")
mrd = DiagrammeRsvg::export_svg(mrd)
mrd = charToRaw(mrd) # flatten
rsvg::rsvg_svg(mrd, "../figures+tables/mrd_bubble.svg")



# *** Global SR map -------------------------------------------------------
rm(list = setdiff(ls(), lsf.str()))  
shp <- readRDS("../processed_data/shp_object_fin_analysis.RDS")
dat_no.na <- readRDS("../processed_data/sem_input_data.RDS")
dat_no.na$level3 <- row.names(dat_no.na)

shp <- shp[shp$LEVEL_3_CO %in% c(dat_no.na$level3),]
rm(dat_no.na)
## get small countries that need a buffer / thicker lines
thicc_lines <- shp[which(shp$area<1.2e+9),]
lcol <- min(thicc_lines$sr)/max(shp$sr)
ucol <- max(thicc_lines$sr)/max(shp$sr)


test <- st_transform(shp, "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs") # Behrmann
test$centroids <- st_centroid(test) %>% 
  st_coordinates()
test$lng <- test$centroids[,1] # x=lng
test$lat <- test$centroids[,2] # y=lat


(sr_map2 <- ggplot(test) + 
    geom_sf(aes(fill=sr),lwd=0, col=NA) + 
    geom_sf(data=thicc_lines, lwd=1.5, aes(col=sr), show.legend=F)+
    scale_colour_viridis_c("SR", option = "plasma", trans = "sqrt", 
                           begin = lcol, end = sqrt(ucol))+
    scale_fill_viridis_c("SR", option = "plasma", trans = "sqrt")+ #, 
    theme(legend.position = c(0.18, 0.3),
          legend.key.height = unit(6,"mm"),
          legend.background = element_blank(),
          legend.key = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          plot.margin = margin(0, 0, 0, 0.5, "cm"),
          panel.border = element_blank(),
          text = element_text(size = 7),
          axis.ticks.x = element_line(color="white"),
       axis.text.x = element_text(color="white")
        )+
    coord_sf(expand = F, label_axes = "--E-", ylim=c(-6200000, 8200000))+
    xlab(" ")
)

# build axis transformation to match Behrmann coordinates with cartesian
# coordinates of scatterplot
sf::sf_use_s2(FALSE) 
shp$centroids <- st_centroid(shp) %>% 
  st_coordinates()
shp$lng <- shp$centroids[,1]
shp$lat <- shp$centroids[,2]

library(pracma)
library(scales)
#transformer for axis
cart <- function(x) {
  xx <- x*pi/180
  tmp <- sin(xx)*sec(30)
  return(tmp)}
cart_inv <- function(x) {
  xx <- asin(x*cos(30))
  xx/(pi/180)}
trans_cart <- trans_new(name = "trans_cart",
                        transform = cart,
                        inverse = cart_inv)

test$sr <- test$sr+10 # tweak to adjust for removed axis padding
(sr_ldg_map <- ggplot(shp, aes(lat,sr,col=sr))+
    geom_point(alpha=0.5)+
    scale_color_viridis_c("SR", option = "plasma", trans = "sqrt")+
    scale_y_continuous("SR/1000", trans = "sqrt", 
                       labels = c("0", "5", "10", "15", "20"),
                       breaks = c(0, 5000, 10000, 15000, 20000))+
     scale_x_continuous("", labels = c("40°S","20°S", "0°", "20°N","40°N","60°N", "80°N"),
                        breaks = c(-40.00000, -20.00000, 0, 20.00000,40.00000,60.00000, 80.00000),
                        position="bottom", trans=trans_cart)+
    geom_smooth(col="grey40", lwd=0.5)+
    coord_flip(xlim=c(-60, 82),ylim =c(0, max(shp$sr)+1500),expand = F)+   
#    coord_trans(y="trans_cart")+
    theme( panel.grid.major = element_blank(),
           panel.grid.minor = element_blank(),
           plot.margin = margin(0.5, 0, 0, -0.2, "cm"),
           legend.position = "none", 
           panel.border = element_blank(),
           axis.line.y = element_line(),
           axis.text.y = element_text(hjust = 0.5),
           text = element_text(size = 7),
           axis.text = element_text(size = 7)))
#(SR <- plot_grid(sr_map2, sr_ldg_map, 
#          ncol = 2, nrow = 1, 
#          rel_widths = c(4.1,1),
#          labels = c("A", ""), label_fontface = "plain"))
#ggsave("../figures/sr_map_AIO_behrmann.png", dpi=600, width=10, height=4.32)



# *** Global MRD map ---------------------------------------------------------

lcol.mrd <- (min(thicc_lines$mrd)-min(shp$mrd))/(max(shp$mrd)-min(shp$mrd))
ucol.mrd <- (max(thicc_lines$mrd)-min(shp$mrd))/(max(shp$mrd)-min(shp$mrd))
(mrd_map2 <- ggplot(test) + 
  geom_sf(aes(fill = mrd), lwd=0 , col=NA) + 
  geom_sf(data=thicc_lines, lwd=1.5, aes(col=mrd), show.legend=F)+
  scale_colour_viridis_c(option = "plasma", 
                         begin = lcol.mrd, end = ucol.mrd)+
  scale_fill_viridis_c("MRD", option = "plasma")+ #, 
    theme(legend.position = c(0.18, 0.3),
          legend.key.height = unit(6,"mm"),
          legend.background = element_blank(),
          legend.key = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          plot.margin = margin(0, 0, 0, 0.5, "cm"),
          panel.border = element_blank(),
          text = element_text(size = 7),
          axis.ticks.x = element_line(color="white"),
          axis.text.x = element_text(color="white"), 
          element_line(size=0) 
    )+
    coord_sf(expand = F, label_axes = "--E-", ylim=c(-6200000, 8200000))+
    xlab(" "))
(mrd_ldg_map <- ggplot(shp, aes(lat,mrd,col=mrd))+
    geom_point(alpha=0.5)+
    scale_color_viridis_c("MRD", option = "plasma")+ 
    scale_y_continuous("MRD")+
    scale_x_continuous("", labels = c("40°S","20°S", "0°", "20°N","40°N","60°N", "80°N"),
                       breaks = c(-40.00000, -20.00000, 0, 20.00000,40.00000,60.00000, 80.00000), 
                       position="bottom", trans=trans_cart)+
    geom_smooth(col="grey40", lwd=0.5)+
    coord_flip(xlim =c(-60, 82), 
               ylim =c(min(test$mrd, na.rm = T)-1, max(test$mrd, na.rm = T)+1),expand = F)+
    theme( panel.grid.major = element_blank(),
           panel.grid.minor = element_blank(),
           plot.margin = margin(0.5, 0, 0, -0.2, "cm"),
           legend.position = "none", 
           panel.border = element_blank(),
           axis.line.y = element_line(),
           axis.text.y = element_text(hjust = 0.5),
           text = element_text(size = 7),
           axis.text = element_text(size = 7))
)
#(MRD <- plot_grid(mrd_map2, mrd_ldg_map, 
#          ncol = 2, nrow = 1, 
#          rel_widths = c(4.1,1), labels = c("B", ""),  label_fontface = "plain"))
#ggsave("../figures/mrd_map_AIO.png", dpi=300, width=11.4, height=9.8496, units="cm")

#p<- plot_grid(SR, MRD, nrow=2)
p <- plot_grid(sr_map2, sr_ldg_map, mrd_map2, mrd_ldg_map, 
          rel_widths = c(4.1,1), labels = c("A", "", "B", ""),
          label_fontface = "plain", label_size = 12)
ggsave(plot=p, "../figures/global_maps_178.pdf", width=20, height=14.7,
       units="cm", bg="white")
ggsave("../figures/global_maps_178.png", width=20, height=14.7, 
       units="cm", device="png", bg="white")

#ggsave("../figures/global_maps.png", dpi=300, width=10, height=7, bg="white")
# ggsave("../figures/global_maps_114.png", dpi=150, width=4.488189, height=2.85611811,
#        bg="white", units="in", device = "png")
# ggsave("../figures/global_maps_114.tiff", dpi=300, width=11.4, height=9.8496)
# 

# ggsave("../figures/global_maps.pdf", width=17.8, height=12.46,
#        device = "pdf", units="cm")
# ggsave2(plot = mrd_ldg_map, "../figures/test.pdf", width=17.8, height=12.46,
#        device = "pdf", units="cm")

# p <- ggplot(data=data.frame(x=1,y=5),aes(x,y))+geom_point()+xlab("more Text")+
#   theme(text = element_text(size = 8),
#         axis.text = element_text(size = 8),
#         axis.title.x = element_text(size = 8))
# p <- sr_map2+ sr_ldg_map + mrd_map2+mrd_ldg_map+ 
#   plot_layout(widths = c(5, 1), byrow = T, ncol=2)
ggsave(plot=p, "../figures/test3.pdf", width=10, height=7,
        device = "pdf")
#pdf("../figures/test_other_functon.pdf", width=10, height=12)
#print(p)
#dev.off()

# *** DR map --------------------------------------------------------------
lcol.mdr <- (min(thicc_lines$mdr)-min(shp$mdr))/(max(shp$mdr)-min(shp$mdr))
ucol.mdr <- (max(thicc_lines$mdr)-min(shp$mdr))/(max(shp$mdr)-min(shp$mdr))
(mdr_map2 <- ggplot(test) + 
    geom_sf(aes(fill = mdr), lwd=0.1) + 
    geom_sf(data=thicc_lines, lwd=1.5, aes(col=mdr), show.legend=F,)+
    scale_colour_viridis_c(option = "plasma", 
                           begin = lcol.mdr, end = ucol.mdr, trans="log")+
    scale_fill_viridis_c("DR", option = "plasma", trans="log")+ #, 
    theme(legend.position = c(0.21, 0.3),
          legend.background = element_blank(),
          legend.key = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          plot.margin = margin(0, 0, 0.5, 0.5, "cm"),
          panel.border = element_blank()
    )+
    coord_sf(expand = F, label_axes = "----")+
    xlab(" "))
(mdr_ldg_map <- ggplot(test, aes(lat,mdr,col=mdr))+
    geom_point(alpha=0.5)+
    scale_color_viridis_c("mdr", option = "plasma", trans="log")+ 
    scale_y_continuous("DR", trans="log")+
    scale_x_continuous("", labels = c("80°S","60°S","40°S","20°S", "0°", "20°N","40°N","60°N", "80°N"),
                       breaks = c(-8000000,-6000000,-4000000, -2000000, 0, 2000000,4000000,6000000, 8000000), 
                       position="bottom")+
    geom_smooth(col="grey40", lwd=0.5)+
    coord_flip(xlim =c(-9020048, 8750122), 
               ylim =c(min(test$mdr, na.rm = T)-1, max(test$mdr, na.rm = T)+1),expand = F)+
    theme( panel.grid.major = element_blank(),
           panel.grid.minor = element_blank(),
           plot.margin = margin(0, 0, 0, -0.4, "cm"),
           legend.position = "none", 
           panel.border = element_blank(),
           axis.line.y = element_line(),
           axis.text.y = element_text(hjust = 0.5))
)
mdr <- plot_grid(mdr_map2, mdr_ldg_map, 
                 ncol = 2, nrow = 1, 
                 rel_widths = c(4,1), labels = c("B", ""),  label_fontface = "plain")



# *** Latitudinal patterns SR + MRD ----------------------------------------

shpbins <- test
# scale data to values between 0 and 1 for comparison
shpbins$sr_scaled <- range01(shpbins$sr)
shpbins$mrd_scaled <- range01(shpbins$mrd)

shpbins <- st_drop_geometry(shpbins)
temp <- pivot_longer(shpbins[,c("lat","mrd_scaled", "sr_scaled", "CONTINENT")], cols = c("mrd_scaled", "sr_scaled"))
temp <- as.data.frame(temp)

temp$g3 <- rep(NA, nrow(temp))
temp$g3[grep("AMERICA", temp$CONTINENT)] <- "Americas"
temp$g3[grep("ASIA", temp$CONTINENT)] <- "Australasia"
temp$g3[grep("AFRICA|EUROPE", temp$CONTINENT)] <- "Afrope"
temp$g3 <- factor(temp$g3, levels = c("Americas","Afrope","Australasia"))

(continent_scatterplot_comb <- ggplot(na.omit(temp), aes(lat, value, col=name))+
    geom_point(alpha=0.5)+
    scale_y_continuous("scaled MRD and SR", labels =c("0", "0.25", "0.50", "0.75", "1"))+#,
    scale_x_continuous("latitude")+
    facet_wrap(~g3, ncol = 3, labeller=
                 labeller(g3 = c(Americas = "Northern & Southern America",
                                 Afrope = "Africa & Europe",
                                 Australasia="Australasia, Asia-Temperate & -Tropical")))+
    geom_smooth(method="loess", se=T)+
    coord_flip(ylim = c(0,1))+
    scale_colour_startrek(labels=c("MRD", "SR"), name="")+
    #theme_void()+
    theme(strip.background = element_blank(),
          strip.text.x = element_text(size = 8),
          panel.border = element_blank(),
          panel.grid = element_blank(),
          axis.line.x = element_line(colour = "grey40"),
          axis.ticks = element_line(colour = "grey40"))+
    guides(color=guide_legend(override.aes=list(fill=NA))) # override geom_smooth() default legend key grey background
)
#ggsave("../figures/map_lat_patterns.pdf", width=7.6, height=4)
ggsave("../figures/map_lat_patterns.png", width=7.6, height=4, dpi=300)

(continent_scatterplot_comb <- ggplot(na.omit(temp), aes(lat, value, col=name))+
    geom_point(alpha=0.5)+
    scale_y_continuous("scaled MRD and SR", labels =c("0", "0.25", "0.50", "0.75", "1"))+#,
    scale_x_continuous("latitude")+
    facet_wrap(~g3, ncol = 3, labeller=
                 labeller(g3 = c(Americas = "Northern & Southern America",
                                 Afrope = "Africa & Europe",
                                 Australasia="Australasia, Asia-Temperate & -Tropical")))+
    geom_smooth(method="loess", se=F)+
    coord_flip(ylim = c(0,1))+
    scale_colour_startrek(labels=c("MRD", "SR"), name="")+
    #theme_void()+
    theme(strip.background = element_blank(),
          strip.text.x = element_text(size = 10),
          panel.border = element_blank(),
          panel.grid = element_blank(),
          axis.line.x = element_line(colour = "grey40"),
          axis.ticks = element_line(colour = "grey40"),
          panel.background = element_rect(fill = "transparent"),
          plot.background = element_rect(fill = "transparent", color = NA))+
    guides(color=guide_legend(override.aes=list(fill=NA))) # override geom_smooth() default legend key grey background
)
#ggsave("../figures/map_lat_patterns_alt.pdf", width=7.6, height=4, bg = "transparent")
ggsave("../figures/map_lat_patterns_alt.png", width=7.6, height=4,
       bg = "transparent", dpi=300)


# save region maps
shp_2 = st_shift_longitude(shp)
# ggplot(shp_2[shp_2$CONTINENT %in% c("NORTHERN AMERICA", "SOUTHERN AMERICA"),]) + 
#   geom_sf(lwd=0, fill = "grey30")+
#   theme(panel.grid = element_blank())
# ggsave("figures/americas.png", width=2, height=1.8, units = "in", dpi = 600)
# 
# ggplot(shp[shp$CONTINENT %in% c("AFRICA", "EUROPE"),]) + 
#   geom_sf(lwd=0, fill = "grey30")+
#   theme(panel.grid = element_blank())
# ggsave("figures/afrope.png", width=1.6, height=2, units = "in", dpi = 600)
# 
ggplot(shp_2[shp_2$CONTINENT %in% c("ASIA-TEMPERATE", "ASIA-TROPICAL", "AUSTRALASIA"),]) +
  geom_sf(lwd=0, fill = "grey30")+
  theme(panel.grid = element_blank())
ggsave("../figures/asia.png", width=2, height=2, units = "in", dpi = 600)

ggplot(shp) + 
  geom_sf(lwd=0, fill = "grey50")+
  coord_sf(expand = FALSE)+
  theme(panel.grid = element_blank(),
        panel.border = element_blank())+
  geom_sf(data=shp_2[shp_2$CONTINENT %in% c("ASIA-TEMPERATE", "ASIA-TROPICAL", "AUSTRALASIA"),], lwd=0, fill = "grey90")+
  theme(panel.grid = element_blank())
ggsave("../figures/world2.png", width=5, height=3, units = "in", dpi = 600)



## latitudinal stats
cor.test(abs(test$lat), test$sr, method = "p") # pearson correlation = -0.38
cor.test(test$lat[test$lat<0], test$sr[test$lat<0], method = "p")
cor.test(test$lat[test$lat>0], test$sr[test$lat>0], method = "p")

cor.test(abs(test$lat), test$mrd, method = "p")
cor.test(test$lat[test$lat<0], test$mrd[test$lat<0], method = "p")
cor.test(test$lat[test$lat>0], test$mrd[test$lat>0], method = "p")
summary(lm(mrd~lat, data=test[test$lat<0,]))
summary(lm(mrd~lat, data=test[test$lat>0,]))
summary(lm(mrd~abs(lat), data=test))


# *** Lat patterns robustness ---------------------------------------------
## bootstrap for different sample size differs between hemispheres 
table(test$lat<0) 
reps <- 500
samp.size <- 50
res.hemisphere <- data.frame(cor.nh.mrd=NA, pval.nh.mrd=NA,
                             cor.sh.mrd=NA, pval.sh.mrd=NA,
                             cor.nh.sr=NA, pval.nh.sr=NA,
                             cor.sh.sr=NA, pval.sh.sr=NA,
                             n=c(1:reps))
set.seed(6348957)
for(i in 1:reps){
  nh <- test[test$lat>0,]
  temp <- nh[sample(1:nrow(nh), samp.size, replace = F),]
  res.hemisphere$cor.nh.sr[i] <- cor.test(temp$lat, temp$sr)$estimate
  res.hemisphere$pval.nh.sr[i] <- cor.test(temp$lat, temp$sr)$p.value
  res.hemisphere$cor.nh.mrd[i] <- cor.test(temp$lat, temp$mrd)$estimate
  res.hemisphere$pval.nh.mrd[i] <- cor.test(temp$lat, temp$mrd)$p.value
  
  sh <- test[test$lat<0,]
  temp <- sh[sample(1:nrow(sh), samp.size, replace = F),]
  res.hemisphere$cor.sh.sr[i] <- cor.test(temp$lat, temp$sr)$estimate
  res.hemisphere$pval.sh.sr[i] <- cor.test(temp$lat, temp$sr)$p.value
  res.hemisphere$cor.sh.mrd[i] <- cor.test(temp$lat, temp$mrd)$estimate
  res.hemisphere$pval.sh.mrd[i] <- cor.test(temp$lat, temp$mrd)$p.value
  
  if(!i%%1)cat(i,"\r")
}
# barplots 
pres <- pivot_longer(res.hemisphere[,c(grep("cor|^n", names(res.hemisphere)))], 
                     cols=contains(c("cor")), names_to = "name", values_to = "cor")
pres2 <- pivot_longer(res.hemisphere[,c(grep("pval|^n", names(res.hemisphere)))], 
                      cols=contains(c("pval")), names_to = "name", values_to = "pval")
pres$name <- gsub("cor.", "", pres$name)
pres2$name <- gsub("pval.", "", pres2$name)
pres3 <- merge(pres,pres2,all=TRUE)
rm(pres,pres2)

pres3$sig <- pres3$pval<0.05
aggregate(cor~name,data=pres3,mean)
aggregate(pval~name,data=pres3,mean)

hemis=
  labeller(name = c(nh.mrd = "northern hemisphere MRD",
                    nh.sr = "northern hemisphere SR",
                    sh.mrd = "southern hemisphere MRD",
                    sh.sr = "southern hemisphere SR"))
ggplot(pres3, aes(cor, fill=sig))+
  geom_histogram()+
  geom_vline(data=ddply(pres3, "name", summarize, cormean = mean(cor)), aes(xintercept=cormean),
             lty=2)+
  geom_vline(xintercept=0)+
  facet_wrap(~name, labeller = hemis)+ #, 
  theme(strip.background = element_blank(),
        strip.text.x = element_text(size = 8))+
  scale_fill_startrek(name="Pvalue < 0.05")+
  xlab("correlation")
  
ggsave("../figures/lat_patterns_robustness.png", width=6, height=4.5, dpi=300)
